revert: drop torch_dtype kwarg from LLaVA load_pretrained_model calls (builder has no **kwargs; original fp16 is the version that worked)

+++ experiments/eval/object_hallucination_vqa_llava.py
@@ -23,7 +23,7 @@ def eval_model(args):
     disable_torch_init()
     model_path = os.path.expanduser(args.model_path)
     model_name = get_model_name_from_path(model_path)
-    tokenizer, model, image_processor, context_len = load_pretrained_model(model_path, args.model_base, model_name, torch_dtype=torch.bfloat16)
+    tokenizer, model, image_processor, context_len = load_pretrained_model(model_path, args.model_base, model_name)
 
     shield.wrap(model, tokenizer,
         caption_file=args.caption_file,
